+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.17.4
+
+* warnings for some use cases of Closeable